feat: environment variable 'PARSQL_TRACE=1' olarak işaretli haldeyken çalıştırılırsa, trace loglarında oluşturulan sql sorgularının console'da gösterilmesini sağlayacak özellikler paketlere eklendi.

+++ examples/postgres/src/select_sample.rs
@@ -8,7 +8,7 @@ use postgres::{types::ToSql, Row, Error};
 #[table("users")]
 #[where_clause("id = $")]
 pub struct SelectUser {
-    pub id: i64,
+    pub id: i32,
     pub name: String,
     pub email: String,
     pub state: i16,
@@ -21,7 +21,7 @@ pub struct SelectUser {
 #[join("LEFT JOIN comments ON posts.id = comments.post_id")]
 #[where_clause("users.id = $")]
 pub struct SelectUserWithPosts {
-    pub id: i64,
+    pub id: i32,
     pub name: String,
     pub email: String,
     pub user_state: i16,
@@ -32,7 +32,7 @@ pub struct SelectUserWithPosts {
 }
 
 impl SelectUserWithPosts {
-    pub fn new(id: i64) -> Self {
+    pub fn new(id: i32) -> Self {
         Self {
             id,
             name: String::default(),
